+++ function/index.js
@@ -2,7 +2,8 @@
 
 const GoogleOAuthAuthorizer = require('./authorization');
 const GoogleDatastoreClient = require('./datastore');
-const { prefixedRejection, ResponseError } = require('./responseError');
+const { prefixedRejection } = require('./responseError');
+const validation = require('./validation.js');
 
 // handle CORS requests via 'cors' library
 const corsOptions = {
@@ -10,91 +11,6 @@ const corsOptions = {
     allowedHeaders: ['Authorization', 'Content-Type', 'Accept', 'Origin', 'X-App-ID'],
 };
 const cors = require('cors')(corsOptions);
-
-function validateRequestUrl(req) {
-    if (req.path !== '/v1/user/response' && req.path !== '/user/response') {
-        throwResponseError(404);
-    }
-}
-
-function validateRequestMethod(req) {
-    if (!['GET', 'POST', 'OPTIONS'].includes(req.method)) {
-        throwResponseError(405);
-    }
-}
-
-function requireAuthorizationHeader(req) {
-    if (req.method === 'GET' || req.method === 'POST') {
-        const authHeader = req.headers.authorization;
-        if (!authHeader) {
-            throwResponseError(401);
-        } else {
-            return authHeader;
-        }
-    }
-}
-
-function validateContentType(req) {
-    if (req.method === 'POST') {
-        const contentTypeHeader = req.headers['content-type'];
-        if (!contentTypeHeader || contentTypeHeader.indexOf('application/json') !== 0) {
-            throwResponseError(415);
-        }
-    }
-}
-
-function validateInputs(req) {
-    let inputErrors = [];
-    let appid, tosversion, accepted;
-
-    if (req.method === 'GET') {
-        appid = req.query['appid'];
-        try {
-            tosversion = parseFloat(req.query['tosversion']);
-        } catch (parseError) {
-            // swallow the error. Unparsable values will result in NaN, and the typecheck
-            // further down in this method (isNaN / typeof != 'number') will handle this case.
-            console.warn('error parsing tosversion value: ' + req.query['tosversion']);
-        }
-    } else if (req.method === 'POST') {
-        if (req.body !== Object(req.body)) {
-            throwResponseError(400, 'Request body must be valid JSON.');
-        }
-        appid = req.body['appid'];
-        tosversion = req.body['tosversion'];
-        accepted = req.body['accepted'];
-        if (typeof accepted !== 'boolean') {
-            inputErrors.push('accepted must be a Boolean.');
-        }
-    } else {
-        // this should never happen, since validateRequestMethod is called before this method
-        throwResponseError(405);
-    }
-
-    if (typeof appid !== 'string' && !(appid instanceof String)) {
-        inputErrors.push('appid must be a String.');
-    }
-    if (isNaN(tosversion) || typeof tosversion !== 'number') {
-        inputErrors.push('tosversion must be a Number.');
-    }
-
-    if (inputErrors.length > 0) {
-        throwResponseError(400, inputErrors.join(' '));
-    } else {
-        let reqinfo = {
-            appid: appid,
-            tosversion: tosversion,
-        };
-        if (typeof accepted !== 'undefined') {
-            reqinfo.accepted = !!accepted;
-        }
-        return reqinfo;
-    }
-}
-
-function throwResponseError(statusCode, message) {
-    throw new ResponseError(message, statusCode);
-}
 
 function respondWithError(res, error, prefix) {
     const code = error.statusCode || 500;
@@ -130,11 +46,11 @@ function tosapi(req, res, authClient, datastoreClient) {
     const authorizer = authClient || new GoogleOAuthAuthorizer();
     const datastore = datastoreClient || new GoogleDatastoreClient();
 
-    validateRequestUrl(req);
-    validateRequestMethod(req);
-    validateContentType(req);
-    const authHeader = requireAuthorizationHeader(req);
-    const reqinfo = validateInputs(req);
+    validation.validateRequestUrl(req);
+    validation.validateRequestMethod(req);
+    validation.validateContentType(req);
+    const authHeader = validation.requireAuthorizationHeader(req);
+    const reqinfo = validation.validateInputs(req);
 
     return authorizer.authorize(authHeader)
         .then(userinfo => {
